chore: setup pages
