+++ src/components/XPcalculator.tsx
@@ -1,0 +1,7 @@
+export const XPcalculator = () => {
+    return (
+        <div>
+            <h1>XP calculator</h1>
+        </div>
+    )
+}
