+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export {BarChart} from './BarChart';
+export {CsvAnalyser} from './CsvAnalyser';
+export {XPcalculator} from './XPcalculator';
